feat: testify and mockery libraries installed, user service mocks moved to own file, assert for testing
