+++ backend/internal/test/user_handler_test.go
@@ -3,72 +3,23 @@ package test
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
-	"log"
 	"net/http"
 	"net/http/httptest"
 	"sigma-test/internal/app"
 	"sigma-test/internal/controller"
 	"sigma-test/internal/entity"
-	"sigma-test/internal/request"
-	"sigma-test/internal/response"
 	"testing"
 
 	"github.com/gin-gonic/gin"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-type MockUserService struct {
-	MockDB []entity.User
-}
-
-func (m *MockUserService) DeleteUser(email string) error {
-	return nil
-}
-
-func (m *MockUserService) GetAllUsers() ([]response.User, error) {
-	return []response.User{}, nil
-}
-
-func (m *MockUserService) GetUserById(id string) (response.User, error) {
-	return response.User{}, nil
-}
-
-func (m *MockUserService) Login(body request.User) (string, error) {
-	return "", nil
-}
-
-func (m *MockUserService) UpdateUser(id string, user request.User) (response.User, error) {
-	return response.User{}, nil
-}
-
-func (m *MockUserService) GetUserByEmail(email string) (response.User, error) {
-	return response.User{}, nil
-}
-
-func (m *MockUserService) CreateUser(user request.User) (response.User, error) {
-	return response.User{}, nil
-}
-
-// SignUp implements controller.UserService.
-func (m *MockUserService) SignUp(body request.User) (response.User, error) {
-	for _, v := range m.MockDB {
-		log.Println(v.Email)
-		log.Println(body.Email)
-
-		if v.Email == body.Email {
-			return response.User{}, errors.New("user already exists")
-		}
-	}
-	return response.User{ID: "test", Email: "test@test.com", Password: "password", Role: "user"}, nil
-}
-
-// Testing signup handler with mocking user service
 func TestSignupWithMocks(t *testing.T) {
 	t.Run("should return 200 on successful signup", func(t *testing.T) {
 		r := gin.New()
 
 		userService := &MockUserService{MockDB: nil}
-
 		controller.InitUserHandler(r, userService)
 
 		requestBody := `{"email": "test@example.com", "password": "password", "role": "user"}`
@@ -76,7 +27,7 @@ func TestSignupWithMocks(t *testing.T) {
 
 		req, err := http.NewRequest("POST", "/api/signup", body)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 		req.Header.Set("Content-Type", "application/json")
 
@@ -84,21 +35,18 @@ func TestSignupWithMocks(t *testing.T) {
 
 		r.ServeHTTP(res, req)
 
-		if res.Code != http.StatusOK {
-			t.Errorf("expected status code %d but got %v", http.StatusOK, res.Code)
-		}
+		assert.Equal(t, http.StatusOK, res.Code)
 	})
 
 	t.Run("should return 400 on invalid body", func(t *testing.T) {
 		r := gin.New()
 
 		userService := &MockUserService{MockDB: nil}
-
 		controller.InitUserHandler(r, userService)
 
 		req, err := http.NewRequest("POST", "/api/signup", nil)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 		req.Header.Set("Content-Type", "application/json")
 
@@ -106,9 +54,7 @@ func TestSignupWithMocks(t *testing.T) {
 
 		r.ServeHTTP(res, req)
 
-		if res.Code != http.StatusBadRequest {
-			t.Errorf("expected status code %d but got %d", http.StatusBadRequest, res.Code)
-		}
+		assert.Equal(t, http.StatusBadRequest, res.Code)
 	})
 
 	t.Run("should return 422 if user exists", func(t *testing.T) {
@@ -117,7 +63,6 @@ func TestSignupWithMocks(t *testing.T) {
 		// Emutaling real users in database
 		mockUser := entity.User{ID: "test", Email: "test@test.com", Password: "test123", Role: "user"}
 		userService := &MockUserService{MockDB: []entity.User{mockUser}}
-
 		controller.InitUserHandler(r, userService)
 
 		requestBody, _ := json.Marshal(mockUser)
@@ -125,21 +70,19 @@ func TestSignupWithMocks(t *testing.T) {
 
 		req, err := http.NewRequest("POST", "/api/signup", body)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 		req.Header.Set("Content-Type", "application/json")
 
 		res := httptest.NewRecorder()
 		r.ServeHTTP(res, req)
 
-		if res.Code != http.StatusUnprocessableEntity {
-			t.Errorf("expected status code %d but got %d", http.StatusUnprocessableEntity, res.Code)
-		}
+		assert.Equal(t, http.StatusUnprocessableEntity, res.Code)
 	})
 }
 
 // Testing signup handler with SetupRouter
-func TestSignup(t *testing.T) {
+func TestSignupIntegration(t *testing.T) {
 	t.Run("should return 200 on successful signup", func(t *testing.T) {
 		router := app.SetupRouter()
 		res := httptest.NewRecorder()
@@ -148,14 +91,12 @@ func TestSignup(t *testing.T) {
 		body := bytes.NewReader(requestBody)
 		req, err := http.NewRequest("POST", "/api/signup", body)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 
 		router.ServeHTTP(res, req)
 
-		if res.Code != http.StatusOK {
-			t.Errorf("expected status code %d but got %d", http.StatusOK, res.Code)
-		}
+		assert.Equal(t, http.StatusOK, res.Code)
 	})
 
 	t.Run("should return 400 on invalid body", func(t *testing.T) {
@@ -164,14 +105,12 @@ func TestSignup(t *testing.T) {
 
 		req, err := http.NewRequest("POST", "/api/signup", nil)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 
 		router.ServeHTTP(res, req)
 
-		if res.Code != http.StatusBadRequest {
-			t.Errorf("expected status code %d but got %d", http.StatusBadRequest, res.Code)
-		}
+		assert.Equal(t, http.StatusBadRequest, res.Code)
 	})
 
 	t.Run("should return 422 if user exists", func(t *testing.T) {
@@ -182,37 +121,27 @@ func TestSignup(t *testing.T) {
 		bodyReader := bytes.NewReader(jsonBody)
 		req, err := http.NewRequest("POST", "/api/signup", bodyReader)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 
 		router.ServeHTTP(res, req)
 
-		// Check the response status code
-		if res.Code != http.StatusUnprocessableEntity {
-			t.Errorf("expected status code %d but got %d", http.StatusUnprocessableEntity, res.Code)
-		}
+		assert.Equal(t, http.StatusUnprocessableEntity, res.Code)
 	})
 }
 
-func TestPingPong(t *testing.T) {
+func TestPingPongIntegration(t *testing.T) {
 	t.Run("should return ping on pong", func(t *testing.T) {
 		router := app.SetupRouter()
 		res := httptest.NewRecorder()
 		req, err := http.NewRequest("GET", "/ping", nil)
 		if err != nil {
-			t.Fatal(err)
+			require.NoError(t, err)
 		}
 
 		router.ServeHTTP(res, req)
 
-		if res.Code != http.StatusOK {
-			t.Errorf("expected status code %d but got %d", http.StatusOK, res.Code)
-		}
-
-		expectedBody := `{"message":"pong"}`
-		if res.Body.String() != expectedBody {
-			t.Errorf("expected body %s but got %s", expectedBody, res.Body.String())
-		}
+		assert.Equal(t, http.StatusOK, res.Code)
+		assert.Equal(t, `{"message":"pong"}`, res.Body.String())
 	})
-
 }
